+++ dbt/adapters/databricks/constraints.py
@@ -137,6 +137,7 @@ class CheckConstraint(TypedConstraint):
             raise self._render_error([["expression"]])
 
     def _render_suffix(self) -> str:
+        assert self.expression is not None  # Validated in _validate()
         if self.expression[0] != "(" or self.expression[-1] != ")":
             return f"CHECK ({self.expression})"
         else: