+++ tests/unit/macros/adapters/test_persist_docs_macros.py
@@ -154,15 +154,13 @@ class TestPersistDocsMacros(MacroTestBase):
         self, template_bundle, context, relation, mock_model_with_columns
     ):
         context["config"] = Mock()
-        context["config"].get = Mock(return_value="delta")
 
         context["api"] = MagicMock()
         context["api"].Column.get_name = Mock(side_effect=lambda col: col["name"])
 
         context["adapter"] = Mock()
-        context["adapter"].has_dbr_capability = Mock(
-            return_value=True
-        )  # Has comment_on_column capability
+        context["adapter"].resolve_file_format.return_value = "delta"
+        context["adapter"].compare_dbr_version = Mock(return_value=0)  # >= 16.1
         context["adapter"].quote = lambda identifier: f"`{identifier}`"
 
         context["run_query_as"] = Mock()
@@ -195,15 +193,13 @@ class TestPersistDocsMacros(MacroTestBase):
         self, template_bundle, context, relation, mock_model_with_columns
     ):
         context["config"] = Mock()
-        context["config"].get = Mock(return_value="delta")
 
         context["api"] = MagicMock()
         context["api"].Column.get_name = Mock(side_effect=lambda col: col["name"])
 
         context["adapter"] = Mock()
-        context["adapter"].has_dbr_capability = Mock(
-            return_value=False
-        )  # No comment_on_column capability
+        context["adapter"].resolve_file_format.return_value = "delta"
+        context["adapter"].compare_dbr_version = Mock(return_value=-1)  # < 16.1
         context["adapter"].quote = lambda identifier: f"`{identifier}`"
 
         context["run_query_as"] = Mock()
@@ -237,7 +233,7 @@ class TestPersistDocsMacros(MacroTestBase):
         self, template_bundle, context, relation, mock_model_with_columns
     ):
         context["config"] = Mock()
-        context["config"].get = Mock(return_value="parquet")
+        context["adapter"].resolve_file_format.return_value = "parquet"
 
         context["log"] = Mock()
         context["run_query_as"] = Mock()